Slider Page

+++ app/Database/Migrations/2022-08-27-093425_CreateBrandTable.php
@@ -33,7 +33,7 @@ class CreateBrandTable extends Migration
                 'null' => true,
             ],
             'status' => [
-                'type' => 'INT',
+                'type' => 'TINYINT',
                 'constraint' => '1',
                 'default' => '1',
             ],
